PK: Add ECDH structures to the interface
We want to enable alternative implementations at this level and
therefore we need a selector for the implementation too.

Since we need to add a selector anyway we can make a cleaner
change and save some memory by adding the function pointers to a
subcontext info instead of the old mbedtls_pk_info_t.

Also this new PK type is different in purpose from MBEDTLS_PK_ECKEY_DH
(and MBEDTLS_PK_ECKEY). Those ifunctionality is limited to hold a key
(either any EC key or a static ECDH key). The newly added type is there
to handle all operations related to ECDH key exchange wether ephemeral
or static.

These types could potentially be extended to handle the key exchanges
too (instead of just holding the key). We won't do this here because
we only want to add support for alternative implementations for ECDH
for now.