Make check_files.py ignore .svg files
There are some .svg files in the docs directory. They are text files and
some lines have Unix line endings, some of them have Windows.

This is picked up by check_files.py and makes it fail. We mark .svg files
in the docs directory as binary to have them ignored. This is acceptable
because these files are generated and shouldn't be edited manually.

Signed-off-by: Janos Follath <janos.follath@arm.com>