mbedtls_rsa_rsassa_pss_*: improve documentation
Hashes used in RSA-PSS:

1. Hashing the message: happens outside of the function call, the input
is either hashed or raw this is outside of the scope of the standard
2. Encoding the input: from the perspective of the standard this hashes the
raw message, from the perspective of the implementation this is the
input to the function wether hashed or raw.
3. Hash in the Mask Generation Function

Remarks:

- Using the same hash for all is more secure reduces the risk of misalignment
(between the signer and verifier) which might be exploitable when weak
hashes are accepted/used on either side.
- The standard recommends to use the same hash for 2 and 3, but does not
say anything about 1.

Issues with the documentation:

- The comments weren't always clear about the thre hashes involved and
often only mentioned two of them (which two varied from function to
function).
- The documentation was giving the impression that the standard
recommends aligning 2 and 1.

Signed-off-by: Janos Follath <janos.follath@arm.com>