ecp_mod_p384: add comments with proof
The first round of carry reduction can not generate a carry thus the
secound round is not needed. The comments illustrating when the
carry is 1. The reduction is simmetric so the case when the carry is
-1 is similar.
The illustration is trying to calculate the input value starting with
setting the carry to 1 before the second round of the carry reduction.
It calculates backwards and tries to determine the value range of
each word. It ends up with a contradiction that the sum of two values is
more than the sum of their maximum.

Signed-off-by: Janos Follath <janos.follath@arm.com>