Remove USE_PSA from psa-migration/strategy.md
MBED_TLS_USE_PSA_CRYPTO is now always enabled we need to remove
documentation discussing cases when it is disabled.

The goal is not to update the document, only to remove
MBED_TLS_USE_PSA_CRYPTO, while making a minimal local context of the
occurrance up to date and sensible.

Signed-off-by: Janos Follath <janos.follath@arm.com>