Remove psa-migration/psa-legacy-bridges.md
This is an architecture document focusing on how PSA APIs can be mixed
with non-PSA APIs, notably including PK (and in fact, it's mostly about
PK, since we didn't identify work to be done in other areas). It is not
really relevant in 4.0/1.0, where the goals will be different — to do
without low-level legacy APIs.

Signed-off-by: Janos Follath <janos.follath@arm.com>